[FIX] industry_fsm: avoid invalid subtask when customer is required
Steps
=====
- Install module industry_fsm
- Create a task from the kanban view
- In the form view, add a title but no customer
- Open the subtasks tab of the notebook
- Add a subtask and add a title to it
- Save the created parent task
- A pop-up indicate that the customer is missing
- Add a customer and save the task

Issue
=====
A pop-up indicate that the subtasks field is invalid and there is no way
to save the current task with its subtask.

Cause
=====
the module industry_fsm introduces a required=True for the field
"partner_id" when edited from a view in the app Field Service. When a
parent task is created without a "partner_id" id set, it can not be
saved, but subtasks can still be created from it. Those child tasks
should have the "partner_id" set to the same value as the one of their
parent. As it is not set in the parent task, its value will be False for
the child task, and changing it in the parent task (by adding a
customer) will not update the child task. As this field is required when
a task is edited from the app Field Service, the task can not be saved.

Fix
===
Two changes are introduced:
- A dependency to "parent_id.partner_id" is added to the method
"_compute_partner_id" of model project.task allowing to update it if
needed when a "partner_id" is set on the parent task.
- The field "patner_id" is not required anymore in the tree view that
display the subtasks in the form view of a fsm task if the customer is
not set on the parent task. This mixed behavior allows:
    - to avoid "missing field: subtasks" message when trying to save a
    task with subtask but no customer set.
    - to prevent user to remove customer from a subtask in field service
    once it is automatically set.

task-3343423
related-122905

closes odoo/enterprise#43515

X-original-commit: 7ea71edc7e06eea856b3848ea8d8d8ade65ffeb7
Related: odoo/odoo#127088
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>