[FIX] website_twitter_wall: edit description in /twitter_walls
Since [1] when the `description` field of the `website.twitter.wall` was
converted into an HTML field, editing the description of a twitter wall
in the `/twitter_walls` page produces an error.
On that page, the description field is within an `<h5>` tag, but upon
edit, text HTML fields are wrapped inside a `<p>`.
The `lxml` `HTMLParser` transforms the received DOM structure:
```xml
<h5 ...>
    <p>Edited value</p>
</h5>
```
which becomes:
```xml
<div>
    <h5 .../>
    <p>Edited value</p>
</div>
```
But to perform the save, `ir.ui.view`'s `save` looks at the `data-oe-*`
attributes on the saved root element.

This commit converts the `<h5>` to a `<div class="h5">` to avoid the
problem.
It also removes the `title` attribute which now contained HTML content.

[1]: https://github.com/odoo/enterprise/commit/776fa2f3bf5cce4e8a6e643440645f08d17278da

task-3289945

closes odoo/enterprise#41751

X-original-commit: 5dab662e3a2c73ce5ead422f48baa29c5c4e4644
Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Signed-off-by: Benoit Socias (bso) <bso@odoo.com>