[FIX] mrp_workorder: missing scrollbar tablet view
**Summary :**
When viewing a work order in tablet view, you mais miss information as
you're not able to scroll

**Before :**
If you have a **quality step** on the operation, you **cannot** scroll up
 and down meaning vital information may be missed.

**After :**
If you have a **quality step** on the operation, you **can** scroll.

**Additional Note :**
If there's a bunch of instructions, pictures... in the same time as a
document on the QCP, the div of the pdf/google slide shrink to 0px.
It explains why **.o_tablet_document** now has a min-height as well.

opw-3326557

closes odoo/enterprise#41894

X-original-commit: c4d4add9c4d698fb3944d3b28d1dc6ee5c79c563
Signed-off-by: William Henrotin (whe) <whe@odoo.com>