[FIX] web_studio: disallow create a one2many pointing to a non-stored many2one
Before this commit, studio allowed to create a one2many pointing to a not-stored
many2one.
As mentionned on https://github.com/odoo/odoo/pull/121664, this is really not ideal as the ORM,
when resolving the records of the one2many, will return all the records and filter out in python afterwards.
This may hinder the performance overall, especially if the many2one is not searchable.

Since there is no easy way to determine in studio many2ones that are non-stored but searchable, this commit just
prevents to add a one2many to a non-stored many2one.

opw-3265982

closes odoo/enterprise#41817

X-original-commit: ca7106d354de8a4c456ac36a58244d4bd13d0707
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>