[FIX] documents: allow edition in list view
Go to Documents, switch to list view, select a record, try to edit
one of its field in the inspector. Before this commit, the change
wasn't applied. This was due to the presence of the attribute
edit="0" in the arch, which disallows the edition since commit [1],
as this attribute now is correctly used in the model to mark all
fields as readonly, and thus prevent them from being edited.

[1] https://github.com/odoo/odoo/commit/81dda7340fcc88f4d6c1186abf0bcbe5fa4d49dc

opw 3325877

closes odoo/enterprise#41816

X-original-commit: 1f1c4b2354843805a4a99103d8bfbe9e6d62d823
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>