[FIX] project_enterprise: use directly the daterange picker in planned dates
Before this commit, when the user defines a planned date for a task
in the form view, he has to click on `Add end date` button to be able
to select the start and end dates directly in the date picker.

This commit allows to directly select the both dates in the date picker
instead of having the planned date end optional in the form and list
views of task.

closes odoo/enterprise#45131

X-original-commit: dc02744dd0667154bb8a91e341805a125221f9d3
Related: odoo/odoo#130651
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>