[FIX] voip: abort obsolete RPC + fix spinner display
With this commit, if a RPC is pending but the search terms change, the
current RPC is aborted and a new one with the updated search terms is
made instead.

This commit also fixes a bug introduced in
https://github.com/odoo/enterprise/pull/41704: sending two RPCs in a row
could desynchronise the state of the interface, displaying the spinner
when it shouldn't.

Part of task-3343258.

closes odoo/enterprise#41881

X-original-commit: 238f6f9f9da97269a93d78d287ba718b7785fdeb
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>