[IMP] l10n_ke_hr_payroll: add tax deduction card report
This commit allows the user to generate Tax Deduction Card
reports for the employees.
This commit also introduces modifications on the payslip computation
by adding bonus and non-cash benefit inputs, pension contribution and
employer costs on the payslip as well as some modifications on the
payslip template.

task-3280883

Part-of: odoo/enterprise#48597