[FIX] knowledge: allow file downloads in mobile app
Steps to reproduce
==================

- Go to knowledge
- Upload a file using the /file command

Try downloading it using the mobile app
"The odoo mobile apps only supports file downloads using ..."

Cause of the issue
==================

The file is downloaded as a blob. This is only supported in a real
browser

Solution
========

Use the download function from @web/core/network/download.
It should only  be used for file download from odoo, but this is
the case.

opw-3308986

closes odoo/enterprise#41790

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>