[FIX] planning: fix gantt_view progressbar percentage issue
In the planning Gantt view, the progress bar percentage is showing incorrectly.

Issue:
For example, if 17 hours are spent from the allocated 40 hours, the percentage
should be 43%, but instead, it is showing 0%.

Cause:
In the above case, 'row.progressBar.ratio' is returning 42.5. In the current
scenario, it is first divided by 100 and then rounded off. As a result,
it returns 0%. However, we want to round it off first, so its value becomes 43,
and then divide it by 100 to obtain the required percentage.

Fix:
'row.progressBar.ratio' is now rounded off first and then divided by 100 to
get the required percentage.

closes odoo/enterprise#47915

Task: 3346509
X-original-commit: 759eb8ac8d4879bcfd2b8bf402a9f98d3ba5204e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>