[FIX] l10n_be_soda: prevent error in case of missing default_journal_id
Before this commit, in the `_l10n_be_parse_soda_file` function,
`journal_company_id` used the `default_journal_id` from context to
find the existing account. However if `default_journal_id` did not
exist in the context, `journal_company_id` would be `False` and is
not able to find the account, if it existed. It then tries to create
the account, leading to this violation of a SQL constraint and causing
this error:
```psycopg2.errors.UniqueViolation: duplicate key value violates unique
constraint \"account_account_code_company_uniq\"```

opw-3302504

closes odoo/enterprise#42085

X-original-commit: 95956ed1ac6bd0379be5d74f864c226629108b79
Signed-off-by: William André (wan) <wan@odoo.com>