[FIX] web_enterprise: click on background home menu scroll to top
On the home menu, there is a hidden input for the search.
By default, the input is present but not visible (`.visually-hidden`)
and has the focus. When the user clicks elsewhere, the blur event is
triggered, and in the Odoo code we force the input to have the focus
again. So the focus by default scroll the view on the element.
To avoid the bug, we use the focus with the `preventScroll` option to
avoid this behavior.

Steps to reproduce:
* Open Odoo
* Enable a tour on CRM (debug mode if needed)
* Install many Apps to have a scrollbar in the home menu or
  resize the window to have a scrollbar in the home menu
* Scroll down a bit to have the first row (App) half visible
* Click anywhere outside the apps grid (e.g. background).
* Window scroll to the top => BUG

OPW-3200840

closes odoo/enterprise#41827

Signed-off-by: Pierre Paridans (app) <app@odoo.com>