[IMP] l10n_be_hr_payroll_account: add new test
The added test is intended to check that an employee signing
a new contract for a new job inside the company will be correctly adapted.
As it doesn't fill any field on the salary configurator, it also
ensures all required field are correctly pre-filled.
To avoid multiple declarations and keep this test aside from the others,
the setup of the test has been moved to a new common file.
Also changes the auto_install value from True to False for the module
test_l10n_be_hr_payroll_account.

task-2898298

closes odoo/enterprise#29718

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>