[FIX] l10n_cl_edi_stock : correct date for delivery guide
The requested date is done date which does not exist before the delivery is validated. So I have used if condition so that it does not produce error if the user tries to generate the report before validating the delivery.

Steps To Reproduce:
1. Add chile localization and create a sales order and validate the delivery.
2. Generate the Delivery Guide from the print drop down.

Current Behaviour:
The date in the report generated is scheduled date.

Expected Behaviour:
The date in the report should be effective date.

OPW-3288178

closes odoo/enterprise#43249

X-original-commit: 3056e5356c0af7d3afaef16e13fe1d06ba9fa8b1
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Hamza Islam (hisl) <hisl@odoo.com>
Signed-off-by: Josse Colpaert <jco@odoo.com>