[FIX] account_accountant: fix date limit in automatic reconcile cron
When the cron runs to reconcile bank statement lines we have a limit
on the date of statements for performances considerations.
Instead of using the date set on the statement minus 3 months,
we now use the `create_date` of the record.
This way, if the user creates a statement more than 3 months
in the past it will now be taken into account by the cron.

opw-3318732

closes odoo/enterprise#41875

Signed-off-by: Laurent Smet <las@odoo.com>