[FIX] account_accountant: fix date limit in automatic reconcile cron
When the cron runs to reconcile bank statement lines we have a limit
on the date of statements for performances considerations.
Instead of using the date set on the statement minus 3 months,
we now use the `create_date` of the record.
This way, if the user creates a statement more than 3 months
in the past it will now be taken into account by the cron.

opw-3318732

closes odoo/enterprise#41972

X-original-commit: 12659be1e14460f79f941d9a11709d958e50761c
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Claire Bretton <clbr@odoo.com>