[IMP] account_reports: avoid double computations when using equivalent expression setups
Reports batch expressions to evaluate, in order to minimize the number of queries to be ran. To do so, they rely on a grouping key containing various parameters to be used when evaluating a report. All computations with similar parameters are batched.

However, in some cases, it was possible that some of these values, while actually different, would lead to the exact same computation. What happened then is that multiple batches were created, and the same query was ran multiple times. On big databases, this could delay the report for entire seconds.

This occurred for example in the Belgian Balance Sheet, because of line 149 (BE_UNAFFECTED_EARNINGS). This line had two problems that caused the creation of a new batch just for it:

1) Its date_scope was "from_beginning", while all other expressions from the same engine were "strict_range" on that report. The Balance Sheet uses a single date, not a range ; so  'strict_range' and 'from_beginning' will give similar results for it.

2) It did not have a groupby value. All other account_codes lines of the report have groupby='account_id', so their grouping key receives next_groupby='account_id', while line 149 gets next_groupby=None.

This commits introduces a generic solution for both of these problems:

1) All reports not using a date_range will now force the date scopes 'strict_range' and 'normal' to 'from_beginning' in their computation.

2) The engines not using the next_groupby value at all can now be declared in a constant (NO_NEXT_GROUPBY_ENGINES). The engines in this constant will always receive None as their next_groupby.

Because of this standardization, engines can now more often be batched, hence improving the performance.

closes odoo/enterprise#41874

Signed-off-by: John Laterre (jol) <jol@odoo.com>