[FIX] helpdesk: speed up _compute_success_rate + add unit test
When the number of active=True tickets for a given team becomes
big, computing the team's success_rate can take a bit of time.
This is coming from the iteration through search_read's results
that becomes a small bottleneck for larger number of tickets.

This commit fixes this by changing the search_read call + iteration
to a call to read_group + search_count. This gives an
average speed up of 4.24 for a customer database with 150 000 tickets
in total.

This commit also adds a unit test for the computation of
success rate to ensure that the basic cases stay correct after
this commit.

opw-3299378

closes odoo/enterprise#42783

X-original-commit: dc4eccc5311afa7d64d117483cdd63c32a0f034c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>