[FIX] web_studio: navigate in a nested x2m which has a context
have a form view like:
```xml
 <form>
      <field name='product_ids'>
          <form>
              <field name="po2m" context="{'context_key': 'value'}">
                  <form>
                      <field name="display_name" />
                  </form>
              </field>
          </form>
     </field>
 </form>
```

In studio, try to edit the form view of po2m (it has a context).

Before this commit, it crashed, as we wanted to evaluate the field's context, but the parent
record (or more precisely, the datapoints that it had in its data) was not in the scope for the evaluation anymore.

After this commit, we evaluate the parent record's context right away. That way it has all the required data to be
used by an another datapoint down the x2many chain.

This is fix was made in an unrelated fix in 16.3: 9d458f8771f507aa9c386530b03db155448b0323

opw-3179365

closes odoo/enterprise#41880

Signed-off-by: Georis François (fge) <fge@odoo.com>