[FIX] account_accountant: deferred entries generation for 'on_validation' method
Currently `on_validation` and `manual` methods worked the same way: the deferred entry is generated on the last day of a period, and reverted the following day.

Now, we'll keep this way of working for the manual & grouped mode, but not for the on validation mode.

Instead, when validating an expense (e.g. 1000e from 1st Jan until 30st April) we will:
- Totally defer the invoice lines (with a deferred start/end date) on the last day of the invoice's month. (e.g.: move 1000 from the expense account to the deferral account)
- For each period in [start, end], create a deferral entry proportionnaly to the total amount to be differred (e.g.: 250e taken from the deferral account to the expense account).

If there is only one period (e.g. 1st Jan until 31st Janu), it means that we don't need to defer the expense/revenue since the invoice deferral and its deferred entry will be created on the same day and will thus cancel each other.

This commit also adds the name of the document to the reference of the generated entries and their lines for simpler auditing.

opw-3338068

X-original-commit: ed95aff6e1192465c40d612b79a9989d42f65f8e
Part-of: odoo/enterprise#41919