[FIX] account_reports: already generated deferred entries
If the user open the Deferred report and clicks multiple times on the
"Generate entry" button, the entries will be generated again for the
same lines.

It also happens if the user goes from "on validation" mode to
"manually and grouped", the already generated deferred entries will be generated again
after clicking the button.

There is now a filter on already deferred entries. Furthermore, a warning
banner is now also displayed on the report if the report shows already deferred
moves to make it clearer that the generated entry might be different from
the report values.

opw-3338068

closes odoo/enterprise#41919

X-original-commit: 028ce302e08599a2c4ec66c29f7e26ff47abf4e6
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>