[FIX] appointment: improve and fix resource management
Improve and fix flows related to booking resources and its
management following our recent implementation. This commit
changes the following points:

- Allow to duplicate an appointment resource

- Display a selection element for resource when we don't
display the avatar to select a resource.

- Fix an issue where the encodeURI was leading to an error
when going back in the process and selecting a new time slot.
Because we never decoded the url params, they were encoded
multiples times. This was leading to unreadable params in the controller.

For example with the param resource_ids=[1], the encoded value should be
"resource_ids=%5B1%5D" with this value in the url when going back and
selecting a new slot the value was encoded a second times leading to this
value: "resource_ids=%255B1%255D" which was encoding the percentage character.
The decodeURIComponent decode the params receive before re-encoding them at
the end allowing to get a proper parameter value.

-Fix the case where when two resources are shareable, linked together
and with resource_manage_capacity enable. Before the condition was never
fulfilled for this case. So the resources were always markes as available,
the remaining capacity was computed. With a perfect match on the capacity asked,
we didn't check the remaining capacity of the resource either as we expected resources
passed to be avaialble. This was leading to resources selected as the best resource
available even if it was already taken.

For example: Table A and Table B have 4 capacity each, are shareable and linked together.
A user make a reservation and Table A is taken completly. Then a user want to book a the
same time for 4 persons. Before, Table A was marked as available and it's total
remaining capacity was equal to 4 (from the Table B linked resource). Then Table A was
selected as best resource but since it's already used, the user was redirected to slot
selection step when submiting the form.

task-3357113

closes odoo/enterprise#43895

X-original-commit: 92a511c6f9c074db925dc0a9f9ae740da13d95e5
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>