[FIX] quality: fix quality alert multi-edit
This commit fixes multi-editing on Quality Alert when stage_id is not in
the vals.

The issue comes from self.stage_id.done, when self.stage_id is not a
singleton. By simply inversing the order of the check (and so by
checking first that 'stage_id' is in the vals), we ensure that
self.stage_id will be a singleton (as the check is located after the
super.write call).

Steps to reproduce:
 - multi-edit quality alerts without changing the stage

Current behavior:
 - a ValueError is thrown: a singleton is expected

Expected behavior:
 - No error, the user can multi-edit without changing the stage

closes odoo/enterprise#42008

Signed-off-by: William Henrotin (whe) <whe@odoo.com>