[FIX] web_grid: avoid reloading when edit a grid cell
Before this commit, when the user edits a cell, the grid view is reloading,
the problem with that method is the rows could be reordered in the grid view
and so the user cannot edit many cells in a same row at the same time.

This commit avoids reloading the grid view when the grid cells are editing
by the user to be able to quickly edit it without any issues.

task-3349578

X-original-commit: 4989f24405972b553064be593b28795f29362169
Part-of: odoo/enterprise#42513