[FIX] timesheet_grid: avoid reloading the grid view when timer stopped
Before this commit, when the timer is stopped by the user, the grid view
is reloading, the problem by doing that is the rows in the grid view could
be reordered and so the user cannot quickly start another timer in a
specific row, he has to wait the timer is stopped before starting another
timer.

This commit avoids reloading the grid view when a timer is stopped to
allow the user to directly start another timer.

task-3349578

X-original-commit: 25a447d54a382b20989f172d06f3a6dde24476fc
Part-of: odoo/enterprise#42513