[FIX] timesheet_grid: allow to start another timer even if a timer is running
Before this commit, when the user uses a shortcut letter to start a timer,
he cannot use another shortcut while a timer is running.

This commit fixes the issue by allowing to use the shortcut to stop
the running timer or to start another timer (if another timer is
started the timer is automatically stopped).

task-3349578

X-original-commit: 6713664aac46258c0108dbe28de4b2a2330dcd5c
Part-of: odoo/enterprise#42513