[FIX] website_sale: better behavior of subscriptions inside website_sale
Issue:

The issue in this part of the fix, is to make sure we can add or we
cannot add a subscription depending on the subscription we might or not
have in our cart.

Steps to reproduce:

-Create 2 products, one with a monthly subscription and a second one
with a yearly subscription only.
-Go to the eCommerce site, and now add one of them to the cart.
-When going back to the other one we should see a message, and we should
be unable to add it to the cart.
-Now go back to the cart and remove the subscription we added.
-And if we return to the subscription that was blocked, we will see that
even we had our cart empty, it is still blocked.

Note: sometimes the bug is not triggered, so it might seem that it is
working as expected, but at some point the subscription will be blocked
and we can't "unblock" it.

Solution:

We have seen that the products pages are cached, and this provokes that
the subscription products that are the ones that depends on what's
happening in the cart won't get a double check, so we added to the
t-cache to also take into account if the _is_add_to_cart_possible()
has changed for any product in order to update the product status.

opw-3084520

closes odoo/enterprise#42431

X-original-commit: bc627f9b070b28217857d624c1eac698b6923491
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>