[IMP] account_accountant: use account based on partner rank
In most cases, a partner is either a customer or a vendor.
Odoo is able to determine this based on the `customer_rank` and `supplier_rank` fields.
We can be certain when one field is positive, and the other is zero.

In this PR, the appropriate receivable or payable account is used (when there is certainty)
If there is no certainty (partner has both customer & vendor rank), the current behaviour is retained, ie. receivable account for incoming transactions, and payable account for outgoing transactions.

Task-3273231

closes odoo/enterprise#42031

Signed-off-by: Laurent Smet <las@odoo.com>