[FIX] account_online_sync: improve error message and allow to open issue directly
Having the reference of the error in the message was confusing and customers have
trouble providing all informations directly when opening an issue with Odoo.
To speed up the process of handling issues, directly propose when an error
occurs, the possibility to open a new issue with Odoo. This is done using
a redirectWarning that will open a new issue with the summary and description
already filled with all the informations needed by the functional support to
analyze the issue

closes odoo/enterprise#47998

Task: 2823656
X-original-commit: 9ff3237b2cb52bbbd5bd7cbe18f72a643d209367
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Cedric Snauwaert <csn@odoo.com>