[ADD] account_avatax_stock: add line level addresses
Before this commit:
When computing the tax using the avatax service, one document level 'addresses' parameter was sent with the company's location as a 'shipFrom' value and the partner's location as a 'shipTo'. When some products are shipped from different locations, this may impact the calculated tax.

After this commit:
When computing the taxes using avatax in a sale order or an invoice created from a sale order with sale and stock installed, we check where each line is delivered from and if delivered from a different address than the company's, we add line level 'addresses' parameter to the line with the correct 'shipFrom' address. Even though the 'shipTo' is the same since the customer's address won't change, it still has to be added to the line level 'addresses' as per the Avatax docs as it no longer inherits the document level 'addresses'.

task-3279421

closes odoo/enterprise#42142

Signed-off-by: Quentin De Paoli <qdp@odoo.com>