[IMP] web_gantt,** : make error message user friendly
Before this commit error messages in odoo are boring and
not much attractive to user. Those were like odoo is preventing
them from doing something user want to do.

In this commit, we have modified the message to be a more friendly
and humorous tone, making it less tedious and more enjoyable. It
aims to enhance the user experience and ensure that interactions
with any application are both pleasant and informative. As a result,
the messages have been clear, short, easy to understand and informative.

task-3356114

closes odoo/enterprise#42472

Related: odoo/upgrade#4899
Related: odoo/odoo#124820
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Co-authored-by: Kartik Chavda <kcv@odoo.com>