[IMP] l10n_be_soda: add account mapping wizard
Currently we support the import of SODA files for Belgium. These are
XML files containing journal entries related to payroll.
The accounts defined in these files come from the payroll service
provider (social secretariat) and might not be the same as the ones used
in the user's accounting.

To ease the import of these files, we add a wizard where the user can
map the accounts from the SODA file to accounts in Odoo. We save this
mapping for subsequent imports to ease the process for the user, while
still allowing him to change the accounts on every import.

task-3090423

closes odoo/enterprise#42363

Signed-off-by: William André (wan) <wan@odoo.com>