[ADD] hr_holidays_contract_gantt: add bridge module
This commit adds an additional module to make the bridge between
the hr_holidays_gantt and the hr_contract modules as the non-working
days display might vary according to different contracts.
This change has been made in order to be consistent with the
changes made for the calendar in community in the hr_contract module.

task-3060722

closes odoo/enterprise#42125

Related: odoo/odoo#121153
Signed-off-by: Kevin Baptiste <kba@odoo.com>