[FIX] l10n_de_reports: payment export
With a Germany company setup
On the Bank journal:
  - Have the 'Bank Account' set as '1204 Bank'
  - In Incoming Payment tab, set a payment method
outstanding receipts accounts equal to '1204 Bank'
Create an invoice
Register a payment, in the journal entry there should be 2 amls for
accounts:
  - 1204 Bank
  - 1410 Forderungen aus Lieferungen u.Leistung
Go to General Ledger
Export DATEV

Issue: Payment will be missing from accounting_entries.csv

This occurs because we will filter out both aml of the move.
The first one, because the account is equal to the counteraccount
The second one, in the attempt to skip outstanding payment/receipt
lines.

opw-3340830

closes odoo/enterprise#42894

X-original-commit: e8b5ea6cbb8d9724da94c034c62f51e10670cb6a
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>