[IMP] l10n_dk_reports: danish standard saft export
The aim of this task is to implement the Danish SAF-T export.

Context:
The Denmark has introduced a certification for accounting Software in
Denmark.
The Danish SAF-T implementation is one of the requirement in order to be
a certified accounting software.

Documentation:
https://erhvervsstyrelsen.dk/standardkontoplan

Before the commit:
There wasn't any standard way for a Danish company to export a SAF-T
file.

After this commit:
Danish companies can export their data using the new SAF-T button on the
general ledger report button.

Task-id: 3253180
Part-of: odoo/enterprise#42561