[IMP] documents{_spreadsheet}: improve url sharing management
Purpose
=======
Improve the sharing management for documents of type url.
A user shouldn't be able to share a url document.

Specifications
==============
When only url documents are selected, prevent the user from
copying the sharing link or accessing the sharing modal by
displaying a danger notification.

When url documents but also other types of documents are selected
in the group of files to share, display a warning message indicating
that documents of type url will be excluded from the shared content.
A warning banner when sharing from the share modal and a warning
notification when sharing from the inspector.

Task-3234864

Part-of: odoo/enterprise#38882