[FIX] knowledge: adapting existing test to accomodate an old one
This commit removes a test that was not well formatted and would not be
launched by the test modules, since it used a cross_module feature and the
test was not located inside the correct module.

To facilitate the transition we are removing the defective test and adapting
an existing one that did a similar kind of test.
Now the behavior is correctly tested and its test is inside a correct module.

task-3358027

closes odoo/enterprise#42287

X-original-commit: b58422b544b9d207e972edc67e2eba1a72182dbf
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>
Signed-off-by: Josse Thomas (thjo) <thjo@odoo.com>