[FIX] project_enterprise,industry_fsm: avoid using replace in arch
Before this commit, the upgrade of a db to 16.2+ could fail because
`<label for="date_deadline"/>` in the arch of a `project.task` form
view is removed when `industry_fsm` module is installed.
Moreover, when Project app is installed, the recurring bouton is
always next to `date_deadline` field instead of being next to
`planned_date_end` when `project_enterprise` module is installed.

This commit reviews the arch to move the recurring bouton next to
`planned_date_end` field when `project_enterprise` instead of doing
it only when `industry_fsm` module is installed. And this commit
avoid doing any replace instead we will use `attributes` to alter
`attrs` and `move` to move the `date_deadline` in the right place
instead of removing it and reintroduce it where we want.

closes odoo/enterprise#42359

X-original-commit: 61da69ddb6b05ffe7423b588839a4e5e3f259edd
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>