[IMP] hr_appraisal: rework appraisal plan
appraisal plan's cron now creates the appraisals on the day of
next_appraisal_date of the hr.employee. when an appraisal is done
it sets that value on the employee.

Moreover, we set constrain on next_appraisal_date, as we don't want it
to be set in the past.
Now, activation of appraisal plan will trigger recomputation of the
next_appraisal_date.
task 3356662

closes odoo/enterprise#42177

Related: odoo/odoo#124452
Signed-off-by: Kevin Baptiste <kba@odoo.com>