[IMP] sale_subscription: support mandate specific values
The Reserve Bank of India (RBI) issued a directive (amended subsequently
in December 2020 and March 2021) that introduces additional security
measures for recurring payments on India-issued cards. These measures
include:
    - Banks must register cardholders and create an e-mandate through a
      one-time process, using additional factor authentication (AFA)
      like 3D Secure (3DS).
    - Banks must alert cardholders at least 24 hours before charges take
      place and give them the ability to opt out of transactions.
    - Recurring transactions over 15,000 INR (or equivalent in other
      currencies) must go through AFA each time.

After this commit, Stripe will create an e-Mandate for every
Indian-based card newly saved in Odoo with the values of the linked
subscription.

task-3322020

Part-of: odoo/enterprise#42196