[FIX] sign: fix adjust signature size when items are detached
In sign, we call adjustSignatureSize to adapt the signature image's dimensions
to match to the sign item box's. However, it is possible that, in some cases,
the sign items are detached from the page elements inside PDF.js. This causes
the width and height of the sign item to be 0px and the calculation is not
possible. This commit checks if the sign items are detached and accordingly
calls refreshSignItems to attach them back before checking the dimensions.

opw-3272737
opw-3094675
opw-3236508

Part-of: odoo/enterprise#42250