[FIX] web_studio: Record and Many2X components domain
Some components in Studio instanciate the Record component and Many2X field widget component directly
(that is, without field.js).

Before the commits [1] and [2], the domain of those fields were retrieved via the RelationalModel. After
those commits, the domain should be passed individually to each component.

This commit adapts the components in Studio for the domains to be correctly passed.

[1]: https://github.com/odoo/odoo/commit/2ecfed335d469
[2]: https://github.com/odoo/odoo/commit/91111f33ef59df7550ee8b2f987fae528a307e0e

should fix runbot-error-20813

closes odoo/enterprise#42249

X-original-commit: 2b512f0e1d5aa7b994fc13de4b840942c1bf868c
Signed-off-by: Dardenne Florent (dafl) <dafl@odoo.com>