[FIX] l10n_pe_edi: Adapt to change in CDR returned by SUNAT
Across several customer DBs, it has been noticed that the CDR returned
by SUNAT via the getStatusCdr endpoint no longer includes the CDR's
original issue date.

This has been causing confusing tracebacks when sending invoices to
SUNAT and when executing the EDI cron if an invoice requires the CDR
retrieval flow to be used (i.e. the CDR already exists on SUNAT's side,
and we need to retrieve it.)

This commit adapts the CDR retrieval to this change in SUNAT's response,
and also changes the method so that in the future it can gracefully
handle fields missing from the CDR.

closes odoo/enterprise#42321

X-original-commit: 30a3455e0790553ffe628ab5a399020b9438ba30
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>