[FIX] pos_sale_stock_renting: add missing dependencies
Before this commit, the pos_sale_stock_renting module only depended on the
point_of_sale and sale_stock_renting modules, while it explicitly uses the
sale_order_line_id field on pos.order.line defined in the pos_sale module.

This problem can result in tracebacks when removing the sales application,
for example, as the pos_sale_stock_renting module is not removed, but the
pos_sale module is. The problem is easily fixed by adding the missing
dependency.

opw-3346677

closes odoo/enterprise#42238

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>