[FIX] l10n_eu_oss_reports: fix sales XML export for Greece/negative amounts
Currently, the OSS Sales XML export is failing validation
by the tax authorities.

This is due to 3 issues:
1. Greece should be using the "EL" country code
instead of the ISO "GR".
2. Negative amounts cannot be declared in the regular `VatAmount`
node due to a positive constraint. It can only be declared in the
`CorrectionsInfo` that is added (with its required fields) here.
3. Lines (a.k.a. countries) with a tax amount of zero should not be
included in the XML.

A test is also added to cover these cases.

task-3280962

closes odoo/enterprise#40019

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>