[FIX] web_studio: make xmlEditor fully scrollable
Before this commit, the xmlEditor was not scrollable.
This was due in part to what https://github.com/odoo/odoo/pull/124585 fixes.

Consequently, this commit sets the context of the CodeEditor to flex. That way
we know it takes all the available space to render the lines.

closes odoo/enterprise#42367

X-original-commit: 1bd295c828f1640f1642c1491e3342b12ac178a0
Related: odoo/odoo#124656
Signed-off-by: Dardenne Florent (dafl) <dafl@odoo.com>