[FIX] appointment: apply timezone on the day
Steps to reproduce:
-------------------
- create an appointment with the
Pacific/Auckland timezone (UTC +12:00);
- with a public user, register for an appointment
on the website before 12 AM.

(We can see the email sent in the chatter)

Issue:
------
The number of the day used in the template
does not correspond to the timezone.

Solution:
---------
Use `mail_tz` to obtain
the day according to the timezone.

opw-3267279

closes odoo/enterprise#42342

Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>