[FIX] timesheet_grid: give the ticket when timer is started in row
Before this commit, when the user launches the timer via the button
timer on a row and that row has helpdesk.ticket then the timer should
have that timer. Also, when the fields in timer header is edited then
the system cannot find the row since it does not know it was a ticket
set on that row.

This commit adds the helpdesk ticket in the dict returned when a timer
is launched to be sure to have all information when the timer header is
updated or when the timer is stopped.

task-3349578

X-original-commit: 7743400e02f692132a5ea29187cb8dceaf8a3932
Part-of: odoo/enterprise#42513