[FIX] timesheet_grid: avoid doing a write when name changes "/" => ""
Before this commit, when the user starts a new timer, the default name
of the new timesheet is "/", in JS, the timer header component will
change that name to remove `/` and have an empty name. That action
will trigger a write rpc call which is useless since the server will
again set "/" when the name is not set.

This commit checks the changes if it is only the name and the change
is '/' => '' then do nothing otherwise do a write rpc call to update
the timesheet with the timer running. Also, the changes done by the
timer header component will no longer notify the relational model
to avoid reload the record.

task-3349578

closes odoo/enterprise#42513

X-original-commit: 93007938123e45e0382d072ebec41245cbdda0ad
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>