[FIX] knowledge: fix the z-index of the action helper
On mobile devices, the action helper currently appears above the sidebar
which makes some items of the sidebar unreachable. This commit will
address that issue by updating the z-index of the different layers.

Step to reproduce the issue:
1. Go on Knowledge using a mobile phone
2. Delete all articles
3. Open the Knowledge editor
4. Toggle the sidebar
=> The action helper appears above the sidebar

TO BE: The action helper should appear behind the backdrop and the sidebar.

Related: https://github.com/odoo/enterprise/pull/34788

task-3367208

closes odoo/enterprise#42387

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>