[FIX] l10n_pe_reports_book: bad inheritance of custom handler
The custom report handler for "PLE Sales 14.1" was inheriting from the one of the generic tax report, instead of inheriting from its parent, common to all tax reports.

Because of that, the caret options couldn't get computed properly when trying to open the report from the UI, and a javascript error was raised.

closes odoo/enterprise#42437

X-original-commit: 20283c5e8e2d184d9129c5fbc480101a4214fe7d
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>